+++ spoti-friends/src/common/models/User/User.swift
