+++ spoti-friends/src/spotify/models/Artist/Artist.swift
