+++ spoti-friends/src/spotify/models/SpotifyProfile/SpotifyProfile.swift
