+++ spoti-friends/src/spotify/models/SpotifyProfile/SpotifyProfileMock.swift
